+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add check to validate that pipeline id is equal to the file name
 * Update log when pipeline file is invalid; JSON parse error will be displayed
 * Better error log, when something's wrong in a step file
+* Improve logger output of extension logger
 
 ## v1.1.1
 * Update `cloud-sdk-webpack` module to v1.5.6.