+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * Hide error stack traces on log level debug
 * Add check to validate that pipeline id is equal to the file name
 * Logging time values are in local time now
+* Update log when pipeline file is invalid; JSON parse error will be displayed
+* Better error log, when something's wrong in a step file
+* Improve logger output of extension logger
 
 ## v1.1.1
 * Update `cloud-sdk-webpack` module to v1.5.6.