+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Improve error handling if user is not logged in
 * Hide error stack traces on log level debug
 * Add check to validate that pipeline id is equal to the file name
+* Logging time values are in local time now
 * Update log when pipeline file is invalid; JSON parse error will be displayed
 * Better error log, when something's wrong in a step file
 * Improve logger output of extension logger