+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Update log when pipeline file is invalid; JSON parse error will be displayed
 * Better error log, when something's wrong in a step file
 * Improve logger output of extension logger
+* Project dependend console commands are usable in all subdirectories of a project
 
 ## v1.1.1
 * Update `cloud-sdk-webpack` module to v1.5.6.