+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Improve logger output of extension logger
 * Project dependend console commands are usable in all subdirectories of a project
 * Add extension create command
+* Update extension workflow; pipelines don't need to be copied to the global pipeline folder anymore (see documentation)
 
 ## v1.1.1
 * Update `cloud-sdk-webpack` module to v1.5.6.